no more "hashmap2" ➞ renamed as "hashmap"
